+++ collector/vllm/collect_attn.py
@@ -62,7 +62,7 @@ class MockAttentionLayer:
 
 
 # https://github.com/vllm-project/vllm/tree/main/vllm/v1/attention/backends
-# support MHA GQA MQA fp16 tensor and float16/fp8 kv cache
+# support MHA GQA MQA bfloat16 tensor and bfloat16/fp8 kv cache
 
 
 @with_exit_stack
@@ -80,7 +80,7 @@ def run_attention_torch(
 ):
     torch.cuda.set_device(device)
 
-    dtype = torch.float16
+    dtype = torch.bfloat16
     model = os.path.join(os.path.dirname(__file__), "fake_hf_model")
     block_size = 64
 
@@ -346,8 +346,8 @@ def run_attention_torch(
         step = input_len
         op_name = "generation_attention"
 
-    kv_cache_dtype_str = "float16" if not use_fp8_kv_cache else "fp8"
-    dtype_str = "float16"
+    kv_cache_dtype_str = "bfloat16" if not use_fp8_kv_cache else "fp8"
+    dtype_str = "bfloat16"
     kernel_source = f"vllm_{backend_name_str}".lower()
 
     log_perf(
